Simplify process access to current_dir
