+++ src/currentprocess.rs
@@ -19,12 +19,8 @@ use enum_dispatch::enum_dispatch;
 #[cfg(feature = "test")]
 use rand::{thread_rng, Rng};
 
-pub mod cwdsource;
 pub mod filesource;
-mod homethunk;
 pub mod terminalsource;
-
-use cwdsource::*;
 
 /// An abstraction for the current process.
 ///
@@ -61,11 +57,11 @@ use cwdsource::*;
 /// methods are in performance critical loops (except perhaps progress bars -
 /// and even there we should be doing debouncing and managing update rates).
 #[enum_dispatch]
-pub trait CurrentProcess: CurrentDirSource + Debug {}
+pub trait CurrentProcess: Debug {}
 
 /// Allows concrete types for the currentprocess abstraction.
 #[derive(Clone, Debug)]
-#[enum_dispatch(CurrentProcess, CurrentDirSource)]
+#[enum_dispatch(CurrentProcess)]
 pub enum Process {
     OSProcess(OSProcess),
     #[cfg(feature = "test")]
@@ -145,12 +141,46 @@ impl Process {
         }
     }
 
+    pub(crate) fn current_dir(&self) -> io::Result<PathBuf> {
+        match self {
+            Process::OSProcess(_) => env::current_dir(),
+            #[cfg(feature = "test")]
+            Process::TestProcess(p) => Ok(p.cwd.clone()),
+        }
+    }
+
     #[cfg(test)]
     fn id(&self) -> u64 {
         match self {
             Process::OSProcess(_) => std::process::id() as u64,
             #[cfg(feature = "test")]
             Process::TestProcess(p) => p.id,
+        }
+    }
+}
+
+impl home::env::Env for Process {
+    fn home_dir(&self) -> Option<PathBuf> {
+        match self {
+            Process::OSProcess(_) => self.var("HOME").ok().map(|v| v.into()),
+            #[cfg(feature = "test")]
+            Process::TestProcess(_) => home::env::OS_ENV.home_dir(),
+        }
+    }
+
+    fn current_dir(&self) -> Result<PathBuf, io::Error> {
+        match self {
+            Process::OSProcess(_) => self.current_dir(),
+            #[cfg(feature = "test")]
+            Process::TestProcess(_) => home::env::OS_ENV.current_dir(),
+        }
+    }
+
+    fn var_os(&self, key: &str) -> Option<OsString> {
+        match self {
+            Process::OSProcess(_) => self.var_os(key),
+            #[cfg(feature = "test")]
+            Process::TestProcess(_) => self.var_os(key),
         }
     }
 }
