+++ src/utils/utils.rs
@@ -11,7 +11,7 @@ use retry::{retry, OperationResult};
 use sha2::Sha256;
 use url::Url;
 
-use crate::currentprocess::{cwdsource::CurrentDirSource, home_process, process};
+use crate::currentprocess::{home_process, process};
 use crate::errors::*;
 use crate::utils::notifications::Notification;
 use crate::utils::raw;
